pyproject.toml
[build-system]
requires = ["uv_build>=0.9.21,<0.10.0"]
build-backend = "uv_build"

[project]
name = "tawala"
version = "1.4.0"
description = ""
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "christianwhocodes", email = "kevin@tawalabora.space" }]
requires-python = ">=3.12"
dependencies = [
    "christianwhocodes>=1.2.4",
    "django>=6.0",
    "django-browser-reload>=1.21.0",
    "django-phonenumber-field[phonenumberslite]>=8.4.0",
    "django-watchfiles>=1.4.0",
    "python-dotenv>=1.2.1",
]

[project.urls]
homepage = "https://github.com/tawalabora/tawala#readme"
repository = "https://github.com/tawalabora/tawala"
issues = "https://github.com/tawalabora/tawala/issues"

[project.scripts]
tawala = "tawala.management.utils.cli:main"
twl = "tawala.management.utils.cli:main"

[dependency-groups]
dev = [
    "djlint>=1.36.4",
    "pillow>=12.0.0",
    "psycopg[binary,pool]>=3.3.2",
    "vercel>=0.3.6",
]

[tool.ruff]
line-length = 101

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true
